separate api.py file + control image return
